1.1.0

+++ src/NadekoBot/Modules/Searches/Commands/Models/WeatherModels.cs
@@ -23,8 +23,8 @@ namespace NadekoBot.Modules.Searches.Commands.Models
     public class Main
     {
         public double temp { get; set; }
-        public int pressure { get; set; }
-        public int humidity { get; set; }
+        public float pressure { get; set; }
+        public float humidity { get; set; }
         public double temp_min { get; set; }
         public double temp_max { get; set; }
     }
